refactor(backend-cuda): split — extract BackendPagedKv (#8 Phase 4)
Move the second-largest supertrait body out of `cuda/mod.rs`:

- `cuda/paged.rs` (~1138 lines): `SplitKScratch` process-global f32
  partial buffers + `with_split_k_scratch` helper +
  `paged_varlen_split_k_dispatch` (split-K varlen prefill path) +
  `paged_batched_flash_dispatch` (Stage 13a batched paged-decode flash,
  `FERRUM_PAGED_FLASH=1`) + `paged_batched_decode_single_pass` (default
  batched paged-decode) + `impl BackendPagedKv for CudaBackend` body.

All helpers stay private to `cuda::paged`; the trait impl is the
only public surface. Pure mechanical extraction — no behaviour change.

After Phase 4: mod.rs **3402 → 2298 lines (-32%)**. Cumulative since
Phase 0: 4756 → 2298 (-52%). Phase 5 (next): `BackendMoeFused`.
